apps/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
requests==2.32.4
python-dateutil==2.9.0.post0
psycopg[binary]==3.2.9
email-validator==2.2.0
openai==1.104.2
anthropic==0.64.0
cryptography==45.0.6
